Build deno [autogenerated commit]

+++ .deno/transforms/aggregations/sum.ts
@@ -11,7 +11,7 @@ class SumExpr implements AggregationComputer<number> {
     }
 
     get type(): _IType<any> {
-        return Types.bigint;
+        return Types.integer;
     }
 
     createGroup(t: _Transaction): AggregationGroupComputer<number> {
